Refactors CLR types to move some into shared space
