+++ types/src/grantRounds.d.ts
@@ -1,5 +1,6 @@
 import { BigNumberish } from 'ethers';
 import { TokenInfo } from '@uniswap/token-lists';
+import { Contribution, GrantPrediction } from './grants';
 
 // GrantRound object from GrantRoundManager
 export type GrantRound = {
@@ -35,3 +36,33 @@ export type GrantRoundMetadata = {
 };
 export type GrantRoundMetadataStatus = 'resolved' | 'pending' | 'error';
 export type GrantRoundMetadataResolution = Partial<GrantRoundMetadata> & { status: GrantRoundMetadataStatus };
+
+/**
+ * Grant rounds contributions and pot details
+ *
+ * @type GrantRoundContributions
+ * @field {grantRound} grant round address
+ * @field {totalPot} total pot amount in the round
+ * @field {matchingTokenDecimals} the number of decimals used by the rounds matching currency
+ * @field {[contributions]} contributions in that round
+ */
+export type GrantRoundContributions = {
+  grantRound: string;
+  totalPot: number;
+  matchingTokenDecimals: number;
+  contributions: Contribution[];
+};
+
+/**
+ * All CLR results/data for a given GrantRound
+ *
+ * @type GrantRoundContributions
+ * @field {grantRound} grant round address
+ * @field {totalPot} total pot amount in the round
+ * @field {matchingTokenDecimals} the number of decimals used by the rounds matching currency
+ * @field {[contributions]} contributions in that round
+ * @field {[predictions]} predictions for the round
+ */
+export type GrantRoundCLR = GrantRoundContributions & {
+  predictions: GrantPrediction[];
+};
